Config updates for USI changes
This is an example of the new way that USI converters are setup. I have tested the config for the central hub (since I use it frequently in my own saves) but not the others. All of the container parts will require updates as well. Note the removal of ModuleLifeSupport. It has been deprecated and replaced by a VesselModule, so it is no longer required as a module on each part.